web_server/tornado_prj/HealthMonitorSys/requirements.txt
tornado
wtforms
pymysql
aiomysql
requests
redis
wtforms_tornado
wtforms_json
bcrypt
peewee-async==0.6.0a0

PyJWT==1.7.1
aiofiles
sockjs-tornado
pyecharts
mysqlclient
py-ecg-detectors

biosppy
matplotlib
numpy
scipy
nolds
spectrum
pyhrv
opensignalsreader